+++ changelog.md
@@ -10,3 +10,9 @@ v1.0
 <li> Reverse a audio file
 <li> Repeat a audio file
 <li> Merge two files together (append a second file to the end of the first)
+
+v1.1
+----
+<li> Added a merge with gap function. Appends with a 1s gap between tracks
+<li> Added an overlay function to allow two tracks to be play at the same time in the file
+